pyproject.toml
[tool.poetry]
name = "goodheart-labs"
version = "0.2.0"
description = ""
authors = [
  "Rob Gordon <rob.gordon@tone-row.com>",
  "Nathan Young <nathanpmyoung@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
requests = "^2.32.3"
asknews = "^0.9.1"
numpy = "^1.26.0"
openai = "^1.57.4"
python-dotenv = "^1.0.1"
forecasting-tools = "^0.2.23"
markdown = "^3.8"
fastapi = "^0.115.12"
uvicorn = { extras = ["standard"], version = "^0.34.2" }
python-multipart = "^0.0.20"
slack-bolt = "^1.23.0"
aiohttp = "^3.12.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
